requirements.txt
num2words~=0.5.13
git+https://github.com/Rapptz/discord.py
dnspython~=2.6.1
motor~=3.5.1
python-dateutil~=2.8.2
pytz~=2024.1
python-decouple
snowflake-id
requests
aiohttp[speedups]~=3.10.5
sentry_sdk
roblox~=2.0.0
zuid~=0.1.1
pytest
typing_extensions
reactionmenu~=3.1.3
jishaku~=2.5.1
oauth2client~=4.1.3
gspread~=5.10.0
sentry_sdk[pymongo]
uvicorn~=0.30.6
fastapi~=0.112.2
aiohttp-cors
pymongo~=4.8.0
pydantic~=2.8.2
gspread_asyncio
fuzzywuzzy
python-Levenshtein
